feat: added client component auth token and refactor
